# [cluster] Store shards in sorted form