<!--  Thanks for sending a pull request!  Here are some tips for you:
1. If this is your first time, please read our contributor guidelines: https://github.com/m3db/m3/blob/master/CONTRIBUTING.md and developer notes https://github.com/m3db/m3/blob/master/DEVELOPER.md
2. Please prefix the name of the pull request with the component you are updating in the format "[component] Change title" (for example "[dbnode] Support out of order writes") and also label this pull request according to what type of issue you are addressing. Furthermore, if this is a WIP or DRAFT PR, please create a draft PR instead: https://github.blog/2019-02-14-introducing-draft-pull-requests/
3. Ensure you have added or ran the appropriate tests for your PR: read more at https://github.com/m3db/m3/blob/master/DEVELOPER.md#testing-changes
4. Follow the instructions for writing a changelog note: read more at https://github.com/m3db/m3/blob/master/DEVELOPER.md#adding-a-changelog
-->

**What this PR does / why we need it**:
<!--
If you have an issue this change addresses, please add the following details:
*Automatically closes linked issue when PR is merged.
Usage: `Fixes #<issue number>`, or `Fixes (paste link of issue)`.
-->
Pretty much all Shards() API calls in use are either `.All()` or `.Shard()/Contains()`, and `.All()` always re-sorts the backing map to produce a sorted list of shards.

m3dbnode doesn't use it heavily, as it manages routing via it's own shardset wrapper. However, m3aggregator does - and it does it over a contended lock - even though it usually is 5 to 30-40 shards, it gets really expensive fast, especially if it has to be done multiple times for staged placements/diffing.
This change trades slower init via `NewShards()` and `.Add()` (the latter is barely used) for much faster `.All()` calls.
While the slowdown in less frequent calls is 1-5x, the speedup is at least an order of magnitude.

```
name                                    old time/op    new time/op    delta
NewShards/16_shards-12                     590ns ± 1%    1445ns ± 1%  +144.92%  (p=0.000 n=10+10)
NewShards/64_shards-12                    2.27µs ± 1%    6.61µs ± 1%  +190.92%  (p=0.000 n=10+9)
NewShards/256_shards-12                   8.92µs ± 0%   35.67µs ± 1%  +299.99%  (p=0.000 n=9+9)
NewShards/1024_shards-12                  35.3µs ± 1%   194.1µs ± 3%  +449.58%  (p=0.000 n=9+10)
NewShards/4096_shards-12                   147µs ± 1%     946µs ± 1%  +543.48%  (p=0.000 n=9+10)
ShardsAllShards/16_shards-12              1.05µs ± 1%    0.09µs ± 1%   -91.31%  (p=0.000 n=9+10)
ShardsAllShards/64_shards-12              6.75µs ± 2%    0.21µs ± 1%   -96.88%  (p=0.000 n=9+10)
ShardsAllShards/256_shards-12             34.9µs ± 2%     0.7µs ± 1%   -97.92%  (p=0.000 n=10+10)
ShardsAllShards/1024_shards-12             172µs ± 0%       3µs ± 1%   -98.21%  (p=0.000 n=6+10)
ShardsAllShards/4096_shards-12             834µs ± 1%      16µs ± 1%   -98.05%  (p=0.000 n=8+9)
ShardsAllShardIDs/16_shards-12             574ns ± 1%      66ns ± 1%   -88.49%  (p=0.000 n=10+10)
ShardsAllShardIDs/64_shards-12            4.27µs ± 1%    0.22µs ± 0%   -94.90%  (p=0.000 n=10+8)
ShardsAllShardIDs/256_shards-12           22.0µs ± 1%     0.8µs ± 1%   -96.35%  (p=0.000 n=10+10)
ShardsAllShardIDs/1024_shards-12           104µs ± 2%       4µs ± 1%   -96.60%  (p=0.000 n=10+10)
ShardsAllShardIDs/4096_shards-12           484µs ± 1%      16µs ± 1%   -96.60%  (p=0.000 n=10+9)
ShardsAdd/16_shards-12                    1.21µs ± 2%    2.02µs ± 2%   +67.24%  (p=0.000 n=9+10)
ShardsAdd/64_shards-12                    5.35µs ± 0%    9.83µs ± 1%   +83.89%  (p=0.000 n=8+10)
ShardsAdd/256_shards-12                   21.2µs ± 0%    49.4µs ± 1%  +133.47%  (p=0.000 n=8+10)
ShardsAdd/1024_shards-12                  83.7µs ± 0%   246.4µs ± 3%  +194.29%  (p=0.000 n=9+10)
ShardsAdd/4096_shards-12                   349µs ± 1%    1700µs ± 1%  +386.85%  (p=0.000 n=9+9)
ShardsEquals/16_shards-12                 2.49µs ± 1%    0.33µs ± 1%   -86.82%  (p=0.000 n=10+10)
ShardsEquals/64_shards-12                 14.7µs ± 1%     1.1µs ± 1%   -92.32%  (p=0.000 n=9+10)
ShardsEquals/256_shards-12                74.2µs ± 2%     4.4µs ± 2%   -94.03%  (p=0.000 n=10+9)
ShardsEquals/1024_shards-12                359µs ± 1%      18µs ± 0%   -95.00%  (p=0.000 n=9+10)
ShardsEquals/4096_shards-12               1.73ms ± 2%    0.08ms ± 0%   -95.35%  (p=0.000 n=10+8)
ShardsNumShardsForState/16_shards-12       205ns ± 1%      35ns ± 0%   -82.87%  (p=0.000 n=8+7)
ShardsNumShardsForState/64_shards-12       738ns ± 3%     137ns ± 1%   -81.40%  (p=0.000 n=9+9)
ShardsNumShardsForState/256_shards-12     3.22µs ± 2%    0.53µs ± 2%   -83.49%  (p=0.000 n=10+9)
ShardsNumShardsForState/1024_shards-12    13.2µs ± 2%     2.5µs ± 1%   -81.32%  (p=0.000 n=10+10)
ShardsNumShardsForState/4096_shards-12    53.1µs ± 1%    11.3µs ± 2%   -78.79%  (p=0.000 n=9+10)
ShardsShard/16_shards-12                  18.0ns ± 1%    18.7ns ± 1%    +3.65%  (p=0.000 n=9+10)
ShardsShard/64_shards-12                  19.1ns ± 3%    19.6ns ± 2%    +2.50%  (p=0.002 n=9+9)
ShardsShard/256_shards-12                 21.4ns ± 3%    21.3ns ± 2%      ~     (p=0.642 n=10+10)
ShardsShard/1024_shards-12                23.5ns ± 1%    24.0ns ± 1%    +2.30%  (p=0.000 n=10+10)
ShardsShard/4096_shards-12                26.1ns ± 2%    26.7ns ± 2%    +2.19%  (p=0.000 n=10+10)
ShardsContains/16_shards-12               17.3ns ± 3%    17.5ns ± 1%    +1.34%  (p=0.003 n=9+10)
ShardsContains/64_shards-12               18.2ns ± 0%    17.9ns ± 1%    -1.76%  (p=0.000 n=7+10)
ShardsContains/256_shards-12              20.0ns ± 0%    20.8ns ± 2%    +4.22%  (p=0.000 n=6+9)
ShardsContains/1024_shards-12             23.2ns ± 2%    24.0ns ± 2%    +3.23%  (p=0.000 n=10+10)
ShardsContains/4096_shards-12             26.4ns ± 3%    26.2ns ± 1%      ~     (p=0.341 n=10+10)

name                                    old allocs/op  new allocs/op  delta
NewShards/4096_shards-12                    3.00 ± 0%      6.00 ± 0%  +100.00%  (p=0.000 n=10+10)
ShardsAllShards/4096_shards-12              2.00 ± 0%      1.00 ± 0%   -50.00%  (p=0.000 n=10+10)
ShardsAllShardIDs/4096_shards-12            2.00 ± 0%      1.00 ± 0%   -50.00%  (p=0.000 n=10+10)
ShardsAdd/4096_shards-12                     137 ± 0%       155 ± 0%   +13.14%  (p=0.000 n=10+7)
ShardsEquals/4096_shards-12                 4.00 ± 0%      1.00 ± 0%   -75.00%  (p=0.000 n=10+10)
```
NB: benches ran on a busy laptop, so the very short calls have noisy values.

**Special notes for your reviewer**:

**Does this PR introduce a user-facing and/or backwards incompatible change?**:
<!--
If no, just write "NONE" in the release-note block below.
If yes, a release note is required:
Enter your extended release note in the block below. If the PR requires additional action from users switching to the new release, include the string "action required".
-->
```release-note
NONE
```

**Does this PR require updating code package or user-facing documentation?**:
<!--
If no, just write "NONE" in the documentation-note block below.
If yes, describe which documentation you updated.
Note: Any changes that significantly updates how a code package is architectured or functions requires code package documentation updates in the form of updates to the README.md file in that package.
-->
```documentation-note
NONE
```